Allow "Semesters to Show" field to be temporarily blank
This allows backspacing to work as a user might expect; i.e. they can
clear the field and enter a new value. If the user leaves the field
blank and unfocuses the field, it will revert back to its last valid
value.

This was more complicated to implement than expected. An auxiliary state
has to be implemented to handle the case when the field is temporarily
blank (a future developer may be able to address this). This auxiliary
state updates the global state (i.e. calls the dispatch function)
through a useEffect hook.

In addition, due to the limitations of the input[type=number] element,
we need to add a onKeyPress handler to disable key presses for
non-numeric keys, as the onChange handler cannot detect when non-numeric
input is given to the field.